Extracted interface for strategy for missing price and retry scheduler
